+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/IIsDirty.cs
@@ -1,0 +1,7 @@
+﻿namespace TD.WPF.ViewModels
+{
+    public interface IIsDirty
+    {
+        bool IsDirty { get; }
+    }
+}